// apps/web/src/shared/api/generated/model/deliveryReadControllerInTransitListSort.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */

export type DeliveryReadControllerInTransitListSort =
  (typeof DeliveryReadControllerInTransitListSort)[keyof typeof DeliveryReadControllerInTransitListSort];

export const DeliveryReadControllerInTransitListSort = {
  closest_delivery: "closest_delivery",
  newest_orders: "newest_orders",
  oldest_orders: "oldest_orders",
  delayed_first: "delayed_first",
  store: "store",
  service: "service",
  title: "title",
  author: "author",
  price: "price",
} as const;